+++ main.py
@@ -42,7 +42,7 @@ def migrate_data():
             store.battle_big = store.battle_big[
                 ['card_detail_id', 'card_name', 'card_type', 'rarity', 'color', 'secondary_color', 'xp', 'gold',
                  'level', 'edition', 'account', 'opponent', 'created_date', 'match_type', 'format', 'mana_cap',
-                 'ruleset1','ruleset2', 'ruleset3', 'inactive', 'battle_id', 'winner', 'result']]
+                 'ruleset1', 'ruleset2', 'ruleset3', 'inactive', 'battle_id', 'winner', 'result']]
 
         if not store.losing_big.empty:
             store.losing_big.loc[store.losing_big.card_name == 'Zyriel', 'secondary_color'] = 'Black'
@@ -58,6 +58,16 @@ def migrate_data():
                     print("Migrate losing battles (format) " + str(counter) + " / " + str(len(battle_id_to_process)))
                     counter += 1
                     store.losing_big.loc[store.losing_big.battle_id == battle_id, 'format'] = battle_format
+
+            battle_id_to_process = store.losing_big.loc[(store.losing_big['opponent'] == 'DRAW')].battle_id.unique().tolist()
+            counter = 0
+            for battle_id in battle_id_to_process:
+                battle = spl.get_battle(battle_id)
+                account = store.losing_big.loc[store.losing_big.battle_id == battle_id].account.tolist()[0]
+                opponent = battle['player_2'] if battle['player_1'] == account else battle['player_1']
+                store.losing_big.loc[store.losing_big.battle_id == battle_id, 'opponent'] = opponent
+                print("Migrate battles (FIX DRAW BATTLES) " + str(counter) + " / " + str(len(battle_id_to_process)))
+                counter += 1
 
     store_util.save_stores()