+++ CHANGES.md
@@ -4,6 +4,7 @@ All notable changes to this project will be documented in this file.
 ## 0.6.0 WIP
 - FIX: skip zero now only skips when sum of DEC/MERIT/SPS == 0, not >0
 - Migrate data: add opponent to battle log 
+- Migrate data: fix opponent for draw battle 
 - Add card overview page, including clickable from main page.
 - Update/Extend Nemesis page